requirements.txt
kafka-python
snowflake-connector-python
faker
python-dotenv
kafka-python==2.0.2
six>=1.15.0
snowflake-connector-python==3.6.0
pandas==2.1.4
python-dotenv==1.0.0
faker==20.1.0
requests==2.31.0
click==8.1.7
sqlalchemy>=2.0.40
responses>=0.25.7
psycopg2-binary==2.9.9
streamlit==1.28.1
plotly==5.17.0
